#54 - testes create campeonato
